ai-chat-app/requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
